Move the shared prop types in the ui shared directory, to reuse them between browser and content components
