+++ app/ui/browser/views/browser.jsx
@@ -15,7 +15,7 @@ import { connect } from 'react-redux';
 import { ipcRenderer } from '../../../shared/electron';
 
 import Style from '../../shared/style';
-import * as SharedPropTypes from '../model/shared-prop-types';
+import * as SharedPropTypes from '../../shared/model/shared-prop-types';
 import BrowserChrome from './browser-chrome';
 import BrowserContent from './browser-content';
 import DeveloperBar from './developerbar';
